[nginx] Build BoringSSL as shared library
Many Lua modules built for NGINX use FFI to access OpenSSL functions,
but this only works if the OpenSSL library is a shared object.

Because the soname has no version, it shouldn't clobber other uses of
OpenSSL/LibreSSL (unless you install the -dev packages). I tried to
include the library in a subdirectory but couldn't figure out how to get
ldconfig working...